+++ c/filters.ml
@@ -39,7 +39,8 @@ module type Basic = sig
 
   val process : ast -> t Or_error.t
 
-  val fuzz : ?seed:int -> t -> o:Lib.Output.t -> config:Config.Fuzz.t -> t Or_error.t
+  val fuzz :
+    ?seed:int -> t -> o:Lib.Output.t -> config:Config.Fuzz.t -> t Or_error.t
 
   val cvars : t -> Config.C_variables.Map.t
   (** [cvars vast] should return a list of C identifiers corresponding to
@@ -103,11 +104,8 @@ Utils.Filter.Make (struct
     let post = B.postcondition vast in
     {Output.cvars; post}
 
-  let run_fuzz ?(seed : int option) (vast : B.t)
-  (oc : Out_channel.t)
-  ~(o : Lib.Output.t)
-  ~(config : Config.Fuzz.t)
-      : Output.t Or_error.t =
+  let run_fuzz ?(seed : int option) (vast : B.t) (oc : Out_channel.t)
+      ~(o : Lib.Output.t) ~(config : Config.Fuzz.t) : Output.t Or_error.t =
     let open Or_error.Let_syntax in
     let%map fz = B.fuzz ?seed ~o ~config vast in
     B.print oc fz ;
@@ -146,13 +144,14 @@ Utils.Filter.Make (struct
       B.Frontend.load_from_ic ~path:(Utils.Io.In_source.to_string src) ic
     in
     let%bind vast = B.process ast in
-    let f = match aux with
-    | Print output_mode ->
-        run_print output_mode
-    | Delitmus ->
-        run_delitmus
-    | Fuzz {seed; o; config} ->
-        run_fuzz ?seed ~o ~config
+    let f =
+      match aux with
+      | Print output_mode ->
+          run_print output_mode
+      | Delitmus ->
+          run_delitmus
+      | Fuzz {seed; o; config} ->
+          run_fuzz ?seed ~o ~config
     in
     f vast oc
 end)
@@ -177,7 +176,8 @@ Make (struct
 
   let process = Mini_convert.translation_unit
 
-  let fuzz ?(seed : int option) (_ : t) ~(o : Lib.Output.t) ~(config : Config.Fuzz.t) : t Or_error.t =
+  let fuzz ?(seed : int option) (_ : t) ~(o : Lib.Output.t)
+      ~(config : Config.Fuzz.t) : t Or_error.t =
     ignore seed ;
     ignore o ;
     ignore config ;
@@ -232,7 +232,12 @@ Make (struct
 
   let delitmus = Delitmus.run
 
-  let fuzz : ?seed:int -> t -> o:Lib.Output.t -> config:Config.Fuzz.t -> t Or_error.t =
+  let fuzz :
+         ?seed:int
+      -> t
+      -> o:Lib.Output.t
+      -> config:Config.Fuzz.t
+      -> t Or_error.t =
     Fuzzer.run
 
   let postcondition :